requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.2
dj-database-url==0.5.0
Django==4.2.14
django-crispy-forms==2.3
django-phonenumber-field==8.0.0
django-summernote==0.8.20.0
gunicorn==20.1.0
phonenumbers==8.13.42
psycopg2==2.9.9
sqlparse==0.5.1
whitenoise==6.5.0
